+++ plugins/inputs/jaeger/input.go
@@ -102,13 +102,20 @@ func (*Input) SampleMeasurement() []inputs.Measurement {
 	return []inputs.Measurement{&itrace.TraceMeasurement{Name: inputName}}
 }
 
+func (ipt *Input) RegHTTPHandler() {
+	if ipt.Endpoint != "" {
+		// itrace.StartTracingStatistic()
+		http.RegHTTPHandler("POST", ipt.Endpoint, handleJaegerTrace)
+	}
+}
+
 func (ipt *Input) Run() {
 	log = logger.SLogger(inputName)
 	log.Infof("%s input started...", inputName)
 	dkio.FeedEventLog(&dkio.Reporter{Message: "jaeger start ok, ready for collecting metrics.", Logtype: "event"})
 
 	// add calculators
-	afterGather.AppendCalculator(itrace.StatTracingInfo)
+	// afterGather.AppendCalculator(itrace.StatTracingInfo)
 
 	// add filters: the order append in AfterGather is important!!!
 	// add close resource filter
@@ -131,7 +138,7 @@ func (ipt *Input) Run() {
 
 	// start up UDP agent
 	if ipt.Address != "" {
-		itrace.StartTracingStatistic()
+		// itrace.StartTracingStatistic()
 		if err := StartUDPAgent(ipt.Address); err != nil {
 			log.Errorf("%s start UDP agent failed: %s", inputName, err.Error())
 		}
@@ -139,13 +146,6 @@ func (ipt *Input) Run() {
 
 	customerKeys = ipt.CustomerTags
 	tags = ipt.Tags
-}
-
-func (ipt *Input) RegHTTPHandler() {
-	if ipt.Endpoint != "" {
-		itrace.StartTracingStatistic()
-		http.RegHTTPHandler("POST", ipt.Endpoint, handleJaegerTrace)
-	}
 }
 
 func init() { //nolint:gochecknoinits