+++ plugins/inputs/opentelemetry/grpc.go
@@ -1,15 +1,13 @@
-// Package opentelemetry is http
+// Package opentelemetry is GRPC : trace & metric
 package opentelemetry
-
-/*
-	开通端口 接收 grpc 数据
-*/
 
 import (
 	"context"
-	"encoding/json"
+	"fmt"
 	"net"
+	"strings"
 
+	"gitlab.jiagouyun.com/cloudcare-tools/datakit/plugins/inputs/opentelemetry/collector"
 	collectormetricepb "go.opentelemetry.io/proto/otlp/collector/metrics/v1"
 	collectortracepb "go.opentelemetry.io/proto/otlp/collector/trace/v1"
 	"google.golang.org/grpc"
@@ -17,29 +15,28 @@ import (
 )
 
 type otlpGrpcCollector struct {
-	TraceEnable  bool   `toml:"trace_enable"`
-	MetricEnable bool   `toml:"metric_enable"`
-	Addr         string `toml:"addr"`
-	stopFunc     func()
+	TraceEnable     bool   `toml:"trace_enable"`
+	MetricEnable    bool   `toml:"metric_enable"`
+	Addr            string `toml:"addr"`
+	ExpectedHeaders map[string]string
+	stopFunc        func()
 }
 
-func (o *otlpGrpcCollector) run() {
+func (o *otlpGrpcCollector) run(storage *collector.SpansStorage) {
 	ln, err := net.Listen("tcp", o.Addr)
 	if err != nil {
-		l.Fatalf("Failed to get an endpoint: %v", err)
+		l.Errorf("Failed to get an endpoint: %v", err)
 		return
 	}
-
 	srv := grpc.NewServer()
 	if o.TraceEnable {
-		et := &ExportTrace{}
+		et := &ExportTrace{storage: storage}
 		collectortracepb.RegisterTraceServiceServer(srv, et)
 	}
 	if o.MetricEnable {
-		em := &ExportMetric{}
+		em := &ExportMetric{storage: storage}
 		collectormetricepb.RegisterMetricsServiceServer(srv, em)
 	}
-
 	o.stopFunc = srv.Stop
 	_ = srv.Serve(ln)
 }
@@ -52,20 +49,20 @@ func (o *otlpGrpcCollector) stop() {
 
 type ExportTrace struct { //nolint:structcheck,stylecheck
 	collectortracepb.UnimplementedTraceServiceServer
-	// errors      []error
-	// requests    int
-	// mu          sync.RWMutex
-	// storage     []*trace.Span
-	// headers     metadata.MD
-	// exportBlock chan struct{}
+	ExpectedHeaders map[string]string
+	storage         *collector.SpansStorage
 }
 
 func (et *ExportTrace) Export(ctx context.Context, //nolint:structcheck,stylecheck
 	ets *collectortracepb.ExportTraceServiceRequest) (*collectortracepb.ExportTraceServiceResponse, error) {
-	l.Infof(ets.String())
-	// ets.ProtoMessage()
+	md, haveHeader := metadata.FromIncomingContext(ctx)
+	if haveHeader {
+		if !checkHandler(et.ExpectedHeaders, md) {
+			return nil, fmt.Errorf("invalid request haeders or nil headers")
+		}
+	}
 	if rss := ets.GetResourceSpans(); len(rss) > 0 {
-		storage.AddSpans(rss)
+		et.storage.AddSpans(rss)
 	}
 	res := &collectortracepb.ExportTraceServiceResponse{}
 	return res, nil
@@ -73,56 +70,36 @@ func (et *ExportTrace) Export(ctx context.Context, //nolint:structcheck,styleche
 
 type ExportMetric struct { //nolint:structcheck,stylecheck
 	collectormetricepb.UnimplementedMetricsServiceServer
-	// errors      []error
-	// requests    int
-	// mu          sync.RWMutex
-	// storage     []*trace.Span
-	// headers     metadata.MD
-	// exportBlock chan struct{}
+	ExpectedHeaders map[string]string
+	storage         *collector.SpansStorage
 }
 
-func (et *ExportMetric) Export(ctx context.Context, //nolint:structcheck,stylecheck
+func (em *ExportMetric) Export(ctx context.Context, //nolint:structcheck,stylecheck
 	ets *collectormetricepb.ExportMetricsServiceRequest) (*collectormetricepb.ExportMetricsServiceResponse, error) {
-	// header
-	header, b := metadata.FromOutgoingContext(ctx)
-	if b {
-		l.Infof("len =%d", header.Len())
-	}
-	l.Infof(ets.String())
-	bts, err := json.MarshalIndent(ets.GetResourceMetrics(), "    ", "")
-	if err == nil {
-		l.Info(string(bts))
-	}
-	// ets.ProtoMessage()
-	orms := make([]*otelResourceMetric, 0)
-	if rss := ets.ResourceMetrics; len(rss) > 0 {
-		for _, resourceMetrics := range rss {
-			tags := toDatakitTags(resourceMetrics.Resource.Attributes)
-			LibraryMetrics := resourceMetrics.GetInstrumentationLibraryMetrics()
-			for _, libraryMetric := range LibraryMetrics {
-				metrics := libraryMetric.GetMetrics()
-				for _, metrice := range metrics {
-					l.Debugf(metrice.Name)
-					bts, err := json.MarshalIndent(metrice, "\t", "")
-					if err == nil {
-						l.Info(string(bts))
-					}
-					l.Infof("metric string=%s", metrice.String())
-					ps := getData(metrice)
-					for _, p := range ps {
-						orm := &otelResourceMetric{
-							name: metrice.Name, attributes: tags,
-							description: metrice.Description, dataType: p.typeName, startTime: p.startTime,
-							unitTime: p.unitTime, data: p.val,
-						}
-						orms = append(orms, orm)
-						// todo 将 orms 转换成 行协议格式 并发送到IO
-					}
-				}
-			}
+	md, haveHeader := metadata.FromIncomingContext(ctx)
+	if haveHeader {
+		if !checkHandler(em.ExpectedHeaders, md) {
+			return nil, fmt.Errorf("invalid request haeders or nil headers")
 		}
 	}
-	l.Infof("orms len=%d", len(orms))
+	if rss := ets.ResourceMetrics; len(rss) > 0 {
+		orms := em.storage.ToDatakitMetric(rss)
+		em.storage.AddMetric(orms)
+	}
 	res := &collectormetricepb.ExportMetricsServiceResponse{}
 	return res, nil
+}
+
+func checkHandler(headers map[string]string, md metadata.MD) bool {
+	if len(headers) == 0 {
+		return true
+	}
+	for k, v := range headers {
+		strs := md.Get(strings.ToLower(k))
+		mdVal := strings.Join(strs, ",")
+		if mdVal != v {
+			return false
+		}
+	}
+	return true
 }