+++ yuque.sh
@@ -47,6 +47,7 @@ fi
 
 LOGGER_PATH=nul dist/datakit-${os}-amd64/datakit \
 	--ignore demo \
+	--cmd-log stdout \
 	--export-manuals .docs \
 	--man-version "${man_version}" \
 	--TODO "-" && \